% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{collectprov}
\alias{collectprov}
\title{Start provenance collection}
\usage{
collectprov()
}
\description{
This function starts rdt to collect provenance.
It collects provenance at the maximum detail level (3) allowed
by rdt and saves debug files.
}
