% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.check}
\alias{.check}
\title{Check if an activity is a function or operation}
\usage{
.check(prov = prov, nsp = nsp, cmd, df_func = df_func)
}
\arguments{
\item{prov}{Parsed provenance}

\item{nsp}{List with the namespaces of used ontologies}

\item{cmd}{Command (activity) to check}

\item{df_func}{Dataframe with R default functions}
}
\description{
This function increases the detail level of the activities recorded by rdt.
}
