% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_activities}
\alias{.get_activities}
\title{Get activities}
\usage{
.get_activities(
  prov = prov,
  nsp = nsp,
  script_file = script_file,
  df_func = df_func
)
}
\arguments{
\item{prov}{Parsed provenance}

\item{nsp}{List with the namespaces of used ontologies}

\item{script_file}{Script dataframe built with the function process_file}

\item{df_func}{Dataframe with R default functions}
}
\description{
This function gets all the activities recorded by rdt.
It gets the expression (entire command) of the activity, checks if its
function or operation, and gets the exeution flow through the informedBy property from rdt.
}
