% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_args}
\alias{.get_args}
\title{Get the parameters for the function activations}
\usage{
.get_args(prov = prov, activities_df = activities_df, nsp = nsp, rdf = rdf)
}
\arguments{
\item{prov}{Parsed provenance}

\item{nsp}{List with the namespaces of used ontologies}

\item{rdf}{RDF object to write the triples}

\item{activities}{Dataframe containing the prov activities recorded by rdt}
}
\description{
This function first get and load the packages used in the script.
For each function activation, it gets the informed arguments.
It then matches the informed arguments with all the arguments needed for the function execution,
to get a complete list of arguments, since some arguments may have default values (parameters).
The function writes the triples, representing all arguments, its parameter value and
relates them with the corresponding function activation (isInputVarOf).
}
