% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_default_functions}
\alias{.get_default_functions}
\title{Get R default functions.}
\usage{
.get_default_functions()
}
\description{
This function builds a dataframe with functions from R's default packages (base, ggplot2, utils, datasets, graphics, grDevices, methods and stats).
This is done to associate functions and packages that rdt does not record in the provenance json file
but that can be used throughout the script.
}
