% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_entity}
\alias{.get_entity}
\title{Get data (input and output) and write the triple}
\usage{
.get_entity(prov = prov, nsp = nsp, rdf = rdf)
}
\arguments{
\item{prov}{Parsed provenance}

\item{nsp}{List with the namespaces of used ontologies}

\item{rdf}{RDF object to write the triples}
}
\description{
This function checks the entities recorded by rdt.
The function checks whether an entity is produced as an output from the
execution of an activity and also serves as an input to another activity.
In this case, the entity is referred to as a Variable. If it is only produced
as a result of an execution, it is referred to as Output.
}
