% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_expression}
\alias{.get_expression}
\title{Get the entire expression of a command#' It then creates the triples with properties of the function, which script
the function relates to (isSubPlanofPlan), and which package it belongs to (hasPackage).}
\usage{
.get_expression(prov = prov, script_df = script_file, activities = df)
}
\arguments{
\item{prov}{Parsed provenance}

\item{script_df}{Script dataframe built with the function process_file}

\item{activities}{Dataframe containing the prov activities recorded by rdt}
}
\description{
This function gets all the lines corresponding to the execution of an activity
and concatenates into one line.
The entire expression will be saved as rdf:value of each function activation or operation.
}
