% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_functions}
\alias{.get_functions}
\title{Get functions and write the triples}
\usage{
.get_functions(
  prov = prov,
  nsp = nsp,
  activities = activities_df,
  rdf = rdf,
  df_func = df_func,
  script = script_file
)
}
\arguments{
\item{prov}{Parsed provenance}

\item{nsp}{List with the namespaces of used ontologies}

\item{activities}{Dataframe containing the prov activities recorded by rdt}

\item{rdf}{RDF object to write the triples}

\item{df_func}{Dataframe with R default functions}

\item{script}{Script dataframe built with the function process_file}
}
\description{
This function gets all functions recorded by rdt and functions not explicitly
defined by rdt (usually functions from default/base packages).
This function also relates the functions to the packages they belong to
(for default and not default packages).
It then creates the triples with properties of the function, which script
the function relates to (isSubPlanofPlan), and which package it belongs to (hasPackage).
}
