% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_functions_activations}
\alias{.get_functions_activations}
\title{Get function activations and write the triple}
\usage{
.get_functions_activations(activities = activities_df, nsp = nsp, rdf = rdf)
}
\arguments{
\item{activities}{Dataframe containing the prov activities recorded by rdt}

\item{nsp}{List with the namespaces of used ontologies}

\item{rdf}{RDF object to write the triples}
}
\description{
This function checks in the activity dataframe registered by rdt which
activities are functions and how many times each function is activated.
It builds a dataframe with the function activation name, which activity (function activation
or operation) preceded its execution and its relation to the function (isStepOfPlan).
This dataframe turns into a triple that is added to the rdf object.
}
