% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_ops}
\alias{.get_ops}
\title{Get operations and write the triple}
\usage{
.get_ops(activities = activities_df, nsp = nsp, rdf = rdf)
}
\arguments{
\item{activities}{Dataframe containing the prov activities recorded by rdt}

\item{nsp}{List with the namespaces of used ontologies}

\item{rdf}{RDF object to write the triples}
}
\description{
This function checks in the activity dataframe registered by rdt which
activities are operations (assigments tha do not involve function execution, for example).
It builds a dataframe with the operation name, which activity (function activation
or operation) preceded its execution and its type.
This dataframe turns into a triple that is added to the rdf object.
}
