% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_packages}
\alias{.get_packages}
\title{Get packages and write the triple}
\usage{
.get_packages(prov = prov, nsp = nsp, rdf = rdf)
}
\arguments{
\item{prov}{Parsed provenance}

\item{nsp}{List with the namespaces of used ontologies}

\item{rdf}{RDF object to write the triples}
}
\description{
This function checks the packages recorded by rdt.
It builds a dataframe with the package name and its version.
This dataframe turns into a triple that is added to the rdf object.
}
