% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_script}
\alias{.get_script}
\title{Get the executed script and write the triple}
\usage{
.get_script(prov = prov, rdf = rdf, nsp = nsp)
}
\arguments{
\item{prov}{Parsed provenance}

\item{rdf}{RDF object to write the triples}

\item{nsp}{List with the namespaces of used ontologies}
}
\description{
This function get the script recorded by rdt.
It also get a more detailed provenance, including the programming language, its version and operating system
in which it was executed.
The function builds a dataframe with the script name and its details, and then
turn this dataframe into a triple that is added to the rdf object.
}
