% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.get_used}
\alias{.get_used}
\title{Get the used property between activity and entity and write the triple}
\usage{
.get_used(
  prov = prov,
  entities = entities_df,
  activities = activities_df,
  nsp = nsp,
  rdf = rdf
)
}
\arguments{
\item{prov}{Parsed provenance}

\item{entities}{Dataframe containing the prov entities recorded by rdt}

\item{activities}{Dataframe containing the prov activities recorded by rdt}

\item{nsp}{List with the namespaces of used ontologies}

\item{rdf}{RDF object to write the triples}
}
\description{
The function checks which entities are used by which activities
(operations or function activations) and it writes the triples.
}
