% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.process_file}
\alias{.process_file}
\title{Get the contents of the R script.}
\usage{
.process_file(filepath)
}
\arguments{
\item{filepath}{Path to the script file}
}
\description{
This function takes the path of the R script as a parameter and returns a dataframe with the content of the script.
Each line of the datrafame corresponds to one line of the script.
The execution of a command can be spread over more than one line.
}
