% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parsetordt}
\alias{parsetordt}
\title{Parse the prov.json file}
\usage{
parsetordt(rdf_file = "output.ttl")
}
\arguments{
\item{rdf_file}{Path to the output ttl file}
}
\description{
This function parses a prov json file.
}
